+++ src/module/item/bondset/bondsetsheet-v2.ts
@@ -3,6 +3,12 @@ import { VueSheetRenderHelperOptions } from '../../vue/vue-render-helper'
 import { VueItemSheet } from '../../vue/vueitemsheet'
 
 export class BondsetSheetV2 extends VueItemSheet {
+  static get defaultOptions() {
+    return mergeObject(super.defaultOptions, {
+      width: 800,
+    })
+  }
+
   get template() {
     return 'systems/foundry-ironsworn/templates/item/bondsetv2.hbs'
   }